Fixes memory leak in a very dirty way
FIX: There's a memory leak somewhere in channels that can be
demonstrated with this simple consumer.  This leak appears to be in the
receive buffer in the channel layer for websocket connections that are
disconnected but still tracked.  Presumption is that they're still
tracked in the channel layer for messages that may eventually come in;
but since the socket consumer is disconnected, they won't receive these
messages anyhow.

Removes the current channel name from that receive buffer in the channel
layer on disconnect.

This requires a peer fix in channels_redis, seen here as a naive hotfix
in `channels_redis/core.py`.  This peer fix acknowledges that the
channels consumer may have removed the entry from the tracking map.